pyproject.toml
[project]
name = "jira-cli"
version = "1.0.0"
description = "Interactive Jira CLI tool"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "click>=8.0",
    "httpx>=0.25",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "questionary>=2.0",
    "keyring>=24.0",
    "markdown-it-py>=3.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]

[project.scripts]
jira = "jira_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true